Update RiderWindow and WikiRiderWorker
This change was made because the QThread in which WikiRiderWorker runned 
was started at the initialization of RiderWindow and it was never 
finished. Now the worker and his thread are created only when the 
"Ride!" button is clicked and are deleted when the worker finish his 
job.

Also a few docstrings were added.